requirements.txt
# Requirements for DR Agent
fastapi
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
alembic==1.17.2
psycopg2-binary==2.9.11
pgvector==0.4.1
pydantic-settings==2.11.0
pydantic==2.12.5
python-dotenv==1.2.1
datasets>=2.20.0
pydantic[email]

# LangChain ecosystem
langchain==1.1.0
langchain-openai==1.1.0
langchain-text-splitters==1.0.0
sentence-transformers==3.0.1
huggingface-hub>=0.24.0
torch==2.9.1
transformers>=4.43.0
google-adk
langchain-ollama==1.0.0

# Observability
langfuse==3.10.1

# Background jobs
apscheduler==3.11.1

# Development/Testing
pytest==9.0.1
pytest-asyncio==1.3.0
httpx==0.28.1